chore: rename saved playground header file
