+++ x-pack/solutions/search/plugins/search_playground/public/components/saved_playground/saved_playground_header.tsx
